+++ src/main/java/model/com/iisquare/jees/oa/service/RoleService.java
@@ -31,14 +31,10 @@ public class RoleService extends ServiceBase {
 	
 	public RoleService() {}
 	
-	public int getCount() {
-		return roleDao.getCount(null, null, null);
-	}
-	
 	public List<Map<String, Object>> getList(String columns, String orderBy, int page, int pageSize) {
 		String append = null;
 		if(!DPUtil.empty(orderBy)) append = DPUtil.stringConcat(" order by ", orderBy);
-		List<Map<String, Object>> list = roleDao.getPage(columns, null, null, append, page, pageSize);
+		List<Map<String, Object>> list = roleDao.getList(columns, null, new Object[]{}, append, page, pageSize);
 		list = ServiceUtil.fillFields(list, new String[]{"status"}, new Map<?, ?>[]{getStatusMap()}, null);
 		list = ServiceUtil.fillRelations(list, memberDao, new String[]{"create_id", "update_id"}, new String[]{"serial", "name"}, null);
 		return list;
@@ -66,9 +62,11 @@ public class RoleService extends ServiceBase {
 	}
 	
 	public int delete(Object... ids) {
-		int count = roleDao.getCount(new String[]{"parent_id"}, new Object[]{ids}, new String[]{"in"}, null);
+		String idStr = DPUtil.safeImplode(",", ids);
+		if(DPUtil.empty(idStr)) return 0;
+		int count = roleDao.getCount(DPUtil.stringConcat("parent_id in (", idStr, " )"), new Object[]{}, null);
 		if(count > 0) return -1;
-		count = memberDao.getCount(new String[]{"role_id"}, new Object[]{ids}, new String[]{"in"}, null);
+		count = memberDao.getCount(DPUtil.stringConcat("role_id in (", idStr, " )"), new Object[]{}, null);
 		if(count > 0) return -2;
 		return roleDao.deleteByIds(ids);
 	}
